[CAMP-2486]-Updated PR comments
