+++ src/App.js
@@ -21,7 +21,6 @@ import CreditCard from './components/atoms/CreditCard'
 import Heading from './components/atoms/Heading'
 import Callout from './components/atoms/Callout'
 import RadioGroup from './components/molecules/RadioGroup'
-import SetSelectorGroup from './components/molecules/SetSelectorGroup'
 import ButtonGroup from './components/molecules/ButtonGroup'
 import Steps from './components/molecules/Steps'
 import Atoms from './examples/atoms'
@@ -43,9 +42,6 @@ const testRightClick = () => console.log('testRightClick:', 'right click')
 
 const onButtonClick = e => console.log('onButtonClick:', e.target)
 
-const handleOnSetSelectorChange = selected =>
-  console.log('handleOnSetSelectorChange:', selected)
-
 const radioOptions = [
   {
     label: 'option 1',
@@ -60,51 +56,6 @@ const radioOptions = [
     value: 'id3',
   },
 ]
-
-let selectorRadioOptions = [
-  {
-    label: 'xs',
-    value: 'xs',
-  },
-  {
-    label: 'x',
-    value: 'x',
-  },
-  {
-    label: 'm',
-    value: 'm',
-  },
-]
-
-let selectorCheckboxOptions = [
-  {
-    label: 'M',
-    value: 'm',
-  },
-  {
-    label: 'T',
-    value: 't',
-  },
-  {
-    label: 'W',
-    value: 'w',
-  },
-  {
-    label: 'Th',
-    value: 'th',
-  },
-  {
-    label: 'F',
-    value: 'f',
-  },
-  {
-    label: 'S',
-    value: 's',
-  },
-]
-
-const defaultRadioVal = ['x']
-const defaultCheckboxVal = ['m', 'w']
 
 function App() {
   const [switch1Status, setSwitch1Status] = useState(false)
@@ -243,24 +194,6 @@ function App() {
           options={radioOptions}
           onChange={handleOnRadioChange}
         ></RadioGroup>
-      </Card>
-
-      <Card style={{ padding: '20px' }}>
-        <h3 style={{ padding: '20px' }}>SetSelector of type Radio</h3>
-        <SetSelectorGroup
-          options={selectorRadioOptions}
-          type="radio"
-          defaultSelected={defaultRadioVal}
-          onChange={handleOnSetSelectorChange}
-        ></SetSelectorGroup>
-
-        <h3 style={{ padding: '20px' }}>SetSelector of type Checkbox</h3>
-        <SetSelectorGroup
-          options={selectorCheckboxOptions}
-          type="checkbox"
-          defaultSelected={defaultCheckboxVal}
-          onChange={handleOnSetSelectorChange}
-        ></SetSelectorGroup>
       </Card>
 
       <Card style={{ padding: '20px' }}>
