+++ src/examples/molecules/index.js
@@ -1,4 +1,5 @@
 import React, { Fragment } from 'react'
+import SetSelectorGroupExample from './SetSelectorGroup'
 import Tooltip from './Tooltip'
 import FlasMessage from './FlashMessage'
 import Modal from './Modal'
@@ -8,5 +9,6 @@ export default () => (
     <Tooltip />
     <FlasMessage />
     <Modal />
+    <SetSelectorGroupExample />
   </Fragment>
 )
